backend/pyproject.toml
[tool.poetry]
name = "saas-ecommerce"
version = "0.1.0"
description = ""
authors = ["Naqib Hakimi <naqib.hakimi.cgi@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
django = "^4.1.5"
graphene-django = "^3.0.0"
django-filter = "^22.1"
channels = "^4.0.0"
daphne = "^4.0.0"
channels-redis = "^4.0.0"
reactivex = "^4.0.4"
pillow = "^9.4.0"
django-graphql-jwt = "^0.3.4"
cloudflare = "^2.11.1"
dnspython = "^2.3.0"
graphene-django-optimizer = "^0.9.1"
django-ratelimit = "^4.0.0"
graphene-file-upload = "^1.3.0"
mypy = "^1.0.0"
twisted = {extras = ["http2", "tls"], version = "^22.10.0"}
django-environ = "^0.9.0"


[tool.poetry.dev-dependencies]
autopep8 = "^2.0.1"

[tool.poetry.group.dev.dependencies]
sourcery-cli = "^1.0.5"
black = "^23.1.0"
flake8 = "^6.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
